requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-address==0.2.8
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-widget-tweaks==1.4.12
idna==3.4
phonenumbers==8.12.57
psycopg2-binary==2.9.4
python-decouple==3.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12